guptonMarrs: Dialog: Don't move dialog to top on mousedown
Change the behavior of the Dialog widget so a dialog no longer moves in front
of other dialogs on the mousedown event. This is because our non-modal dialogs
(popups) dismiss when you click away, and if the click-away is on another
dialog, we don't want the other dialog's z-index to change simply because
you're dismissing a popup.